+++ ft_putstr.c
@@ -6,7 +6,7 @@
 /*   By: lasalmi <lasalmi@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/03 15:11:04 by lasalmi           #+#    #+#             */
-/*   Updated: 2021/12/02 12:31:36 by lasalmi          ###   ########.fr       */
+/*   Updated: 2021/12/05 13:19:12 by lasalmi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr(char const *s)
 {
-	int	bytes;
+	size_t	bytes;
 
 	if (!s)
 		return ;
